# RegulAgentV2/regulagent-backend/ra_config/openapi.yaml
openapi: 3.0.3
info:
  title: RegulAgent API
  version: "0.1.0"
  description: |
    Core endpoints for plan generation, artifacts, agentic modifications, and similar wells.
servers:
  - url: http://127.0.0.1:8001
    description: Host machine (dev proxy)
  - url: http://localhost:8000
    description: Inside Docker web container
...
paths:
...
  /api/similar-wells:
    get:
      tags: [SimilarWells]
      summary: Find similar wells (geo + structured similarity; optional vector rerank)
      parameters:
        - in: query
          name: api
          required: true
          schema: { type: string }
        - in: query
          name: radius_mi
          schema: { type: number, default: 50 }
        - in: query
          name: k
          schema: { type: integer, default: 20 }
        - in: query
          name: months
          schema: { type: integer, default: 24 }
        - in: query
          name: operator_boost
          schema: { type: number, default: 1.2 }
        - in: query
          name: field_boost
          schema: { type: number, default: 1.1 }
        - in: query
          name: w_architecture
          schema: { type: number, default: 0.4 }
        - in: query
          name: w_pattern
          schema: { type: number, default: 0.45 }
        - in: query
          name: w_context
          schema: { type: number, default: 0.1 }
        - in: query
          name: w_quality
          schema: { type: number, default: 0.05 }
        - in: query
          name: alpha_geo
          schema: { type: number, default: 0.5 }
        - in: query
          name: use_vector
          schema: { type: boolean, default: false }
        - in: query
          name: vector_weight
          schema: { type: number, default: 0.4 }
        - in: query
          name: require_county
          schema: { type: string, enum: [true, false, auto], default: auto }
        - in: query
          name: require_field
          schema: { type: string, enum: [true, false, auto], default: auto }
        - in: query
          name: county
          schema: { type: string }
        - in: query
          name: field
          schema: { type: string }
        - in: query
          name: operator
          schema: { type: string }
      responses:
        "200":
          description: Similar wells
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarWellsResponse'
        "404":
          description: Source well missing or no lat/lon
...
  /api/public/wells/{id}:
    get:
      tags: [Public]
      summary: Well detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detail
...
components:
  schemas:
...
    SimilarWellsResponse:
      type: object
      properties:
        api: { type: string }
        lat: { type: number }
        lon: { type: number }
        radius_mi: { type: number }
        months: { type: integer }
        operator_boost: { type: number }
        field_boost: { type: number }
        rerank: { type: string, nullable: true }
        vector_weight: { type: number, nullable: true }
        count: { type: integer }
        neighbors:
          type: array
          items:
            type: object
            properties:
              api14: { type: string }
              state: { type: string }
              county: { type: string, nullable: true }
              distance_mi: { type: number }
              operator_name: { type: string, nullable: true }
              field_name: { type: string, nullable: true }
              score: { type: number }
              components:
                type: object
                properties:
                  base_geo: { type: number }
                  arch: { type: number }
                  pattern: { type: number }
                  context: { type: number }
              vector_sim: { type: number, nullable: true }
              blended_score: { type: number, nullable: true }